refactor: introduce shared barrels for ast and utils
