+++ src/cli/lib/env-overrides.js
@@ -1,7 +1,6 @@
 import { CliUsageError } from "./cli-errors.js";
-import { isNonEmptyString } from "../../shared/string-utils.js";
+import { isNonEmptyString, isObjectLike } from "../../shared/utils.js";
 import { isErrorLike } from "../../shared/utils/capability-probes.js";
-import { isObjectLike } from "../../shared/object-utils.js";
 
 const DEFAULT_SOURCE = "env";
