+++ src/cli/lib/manual-utils.js
@@ -1,8 +1,10 @@
 import fs from "node:fs";
 import path from "node:path";
-import { parseJsonWithContext } from "../../shared/json-utils.js";
-import { toTrimmedString } from "../../shared/string-utils.js";
-import { formatBytes } from "../../shared/number-utils.js";
+import {
+    formatBytes,
+    parseJsonWithContext,
+    toTrimmedString
+} from "../../shared/utils.js";
 import { ensureDir } from "./file-system.js";
 import { formatDuration } from "./time-utils.js";
