+++ src/cli/lib/vm-eval-timeout.js
@@ -1,7 +1,7 @@
 import {
     coerceNonNegativeInteger,
     resolveIntegerOption
-} from "../../shared/numeric-option-utils.js";
+} from "../../shared/utils.js";
 
 export const DEFAULT_VM_EVAL_TIMEOUT_MS = 5000;
