+++ src/parser/src/gml-syntax-error.js
@@ -1,6 +1,6 @@
 import antlr4 from "antlr4";
 
-import { isNonEmptyString } from "../../shared/string-utils.js";
+import { isNonEmptyString } from "../../shared/utils.js";
 
 const { ErrorListener } = antlr4.error;
