+++ src/shared/ast.js
@@ -1,0 +1,1 @@
+export * from "./ast/index.js";
